+++ SharedProject1/Commands/Command2.cs
@@ -1,21 +1,17 @@
 ﻿using Microsoft.VisualStudio.Shell;
+using Recoding.ClippyVSPackage;
 using System;
 using System.ComponentModel.Design;
-using System.Globalization;
-using System.Windows.Forms;
-using Task = System.Threading.Tasks.Task;
+using AsyncTasks = System.Threading.Tasks;
 
-namespace Recoding.ClippyVSPackage
+namespace SharedProject1
 {
-    /// <summary>
-    /// Command handler
-    /// </summary>
-    internal sealed class Command1
+    internal sealed class Command2
     {
         /// <summary>
         /// Command ID.
         /// </summary>
-        private const int CommandId = 4129;
+        private const int CommandId = 4130;
 
         /// <summary>
         /// Command menu group (command set GUID).
@@ -23,32 +19,41 @@ namespace Recoding.ClippyVSPackage
         private static readonly Guid CommandSet = new Guid("fbed79a9-1faa-4dc3-9f96-9fb39d31bfdb");
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="Command1"/> class.
+        /// VS Package that provides this command, not null.
+        /// </summary>
+        private readonly AsyncPackage _package;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Command2"/> class.
         /// Adds our command handlers for menu (commands must exist in the command table file)
         /// </summary>
+        /// <param name="package">Owner package, not null.</param>
         /// <param name="commandService">Command service to add command to, not null.</param>
-        private Command1(OleMenuCommandService commandService)
+        private Command2(AsyncPackage package, OleMenuCommandService commandService)
         {
+            _package = package ?? throw new ArgumentNullException(nameof(package));
             commandService = commandService ?? throw new ArgumentNullException(nameof(commandService));
 
             var menuCommandId = new CommandID(CommandSet, CommandId);
-            var menuItem = new MenuCommand(Execute, menuCommandId);
+            var menuItem = new MenuCommand(this.Execute, menuCommandId);
             commandService.AddCommand(menuItem);
         }
+
 
         /// <summary>
         /// Initializes the singleton instance of the command.
         /// </summary>
         /// <param name="package">Owner package, not null.</param>
-        public static async Task InitializeAsync(AsyncPackage package)
+        public static async AsyncTasks.Task InitializeAsync(AsyncPackage package)
         {
-            // Switch to the main thread - the call to AddCommand in Command1's constructor requires
+            // Switch to the main thread - the call to AddCommand in Command2's constructor requires
             // the UI thread.
-            //await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(package.DisposalToken);
+            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(package.DisposalToken);
 
             OleMenuCommandService commandService = await package.GetServiceAsync(typeof(IMenuCommandService)) as OleMenuCommandService;
-            _ = new Command1(commandService);
+            _ = new Command2(package, commandService);
         }
+
 
         /// <summary>
         /// This function is the callback used to execute the command when the menu item is clicked.
@@ -60,11 +65,8 @@ namespace Recoding.ClippyVSPackage
         private void Execute(object sender, EventArgs e)
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-            string message = string.Format(CultureInfo.CurrentCulture, "Inside {0}.MenuItemCallback()", this.GetType().FullName);
-            string title = "Command1";
-
-            // Show a message box to prove we were here
-            MessageBox.Show(message, title, MessageBoxButtons.OK, MessageBoxIcon.Information);
+            ((ClippyVisualStudioPackage)_package).ReviveMerlinCommand();
         }
     }
+
 }
