+++ SharedProject1/Configurations/ClippySingleAnimation.cs
