+++ SharedProjectTests2/SharedProjectTests2.csproj
@@ -54,7 +54,9 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="UnitTest1.cs" />
+    <Compile Include="ClippyTests.cs" />
+    <Compile Include="MerlinTests.cs" />
+    <Compile Include="GeniusTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
